requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2